# docs/api.yml
openapi: 3.0.0
servers:
  - url: https://www.bhut.com.br
info:
  version: 1.0.0
  title: TIMESHEET API
  description: API for timesheet
  contact:
    name: BHUT
    url: 'https://www.bhut.com.br'
    email: 'contato@bhut.com.br'
  license:
    name: MIT
    url: https://api.github.com/licenses/mit
...
paths:
...
  /v1/user/{id}:
    get:
      tags:
        - User
      summary: 'Get user'
      description: 'Get user details'
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
    put:
      tags:
        - User
      summary: 'Update user'
      description: 'Update an existent user'
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
      requestBody:
        description: 'User data'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '204':
          description: 'Request processed'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
  /v1/user/{id}/password:
    put:
      tags:
        - User
      summary: 'Update user password'
      description: 'Update an existent user password'
      operationId: updateUserPassword
      parameters:
        - name: id
          in: path
          description: 'Identifier to filter'
          schema:
            type: string
            format: uuid
            example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
          required: true
      requestBody:
        description: 'User data'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordRequest'
        required: true
      responses:
        '204':
          description: 'Request processed'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Access denied'
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
        '500':
          description: 'Internal server error'
      security:
        - ApiAuth: [ ]
        - JwtAuth: [ ]
...
components:
  schemas:
...
    ResourceNotFoundResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                enum:
                  - 'RESOURCE_NOT_FOUND'
              message:
                type: string
                example: 'Resource not found'
...
    UpdateUserPasswordRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          description: 'User password'
          example: 'password@123'
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        email:
          type: string
          description: 'User email'
          example: 'email@domain.com'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        hourValue:
          type: number
          description: 'User hour value'
          example: 17.05
        hasBankHours:
          type: boolean
          description: 'Define if user has bank hours'
        contractTotal:
          type: integer
          description: 'User contract total hours'
          example: 176
        groupId:
          type: string
          description: 'User group identifier'
          example: '6de140ca-ae94-4e1d-9c32-8de18fa22e45'
        startDate:
          type: string
          description: 'User start date'
          example: '2022-01-04'
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 'User identifier'
          example: 'd883fed5-49fb-435d-b719-e3ce0c8d7cf0'
        name:
          type: string
          description: 'User name'
          example: 'John Doe'
        email:
          type: string
          description: 'User email'
          example: 'email@domain.com'
        team:
          type: string
          description: 'User team'
          example: 'Bhut team'
        hourValue:
          type: number
          description: 'User hour value'
          example: 17.05
        hasBankHours:
          type: boolean
          description: 'Define if user has bank hours'
          example: true
        totalBankHours:
          type: number
          description: 'User total bank hours'
          example: 9.51
        contractTotal:
          type: integer
          description: 'User contract total hours'
          example: 176
        groupId:
          type: string
          description: 'User group identifier'
          example: '6de140ca-ae94-4e1d-9c32-8de18fa22e45'
        groupName:
          type: string
          description: 'User group name'
          example: 'Developer'
        startDate:
          type: string
          description: 'User start date'
          example: '2022-01-04'
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                enum:
                  - 'INVALID_FIELD'
                  - 'INVALID_HEADER'
              message:
                type: string
                example: 'Invalid/required fields'
  securitySchemes:
    JwtAuth:
      type: controllers
      scheme: bearer
      bearerFormat: JWT
    ApiAuth:
      in: header
      type: apiKey
      name: X-Api-Key